File render-pages.js is created
