+++ src/render-pages.js
@@ -1,0 +1,182 @@
+import { getPosts, addPost, getUserPosts, deletePost } from "./api.js";
+import { renderAddPostPageComponent } from "./components/add-post-page-component.js";
+import { renderAuthPageComponent } from "./components/auth-page-component.js";
+import {
+  ADD_POSTS_PAGE,
+  AUTH_PAGE,
+  LOADING_PAGE,
+  POSTS_PAGE,
+  USER_POSTS_PAGE,
+} from "./routes.js";
+import { renderPostsPageComponent } from "./components/posts-page-component.js";
+import { renderLoadingPageComponent } from "./components/loading-page-component.js";
+import {
+  getUserFromLocalStorage,
+  removeUserFromLocalStorage,
+  saveUserToLocalStorage
+} from "./helpers.js";
+
+export let user = getUserFromLocalStorage();
+export let page = null;
+export let posts = [];
+
+const getToken = () => {
+  const token = user ? `Bearer ${user.token}` : undefined;
+  return token;
+};
+
+export const logout = () => {
+  user = null;
+  removeUserFromLocalStorage();
+  goToPage(POSTS_PAGE);
+};
+
+/**
+ * Включает страницу приложения
+ */
+export const goToPage = (newPage, data) => {
+  if (
+    [
+      POSTS_PAGE,
+      AUTH_PAGE,
+      ADD_POSTS_PAGE,
+      USER_POSTS_PAGE,
+      LOADING_PAGE,
+    ].includes(newPage)
+  ) {
+    if (newPage === ADD_POSTS_PAGE) {
+      // Если пользователь не авторизован, то отправляем его на авторизацию перед добавлением поста
+      page = user ? ADD_POSTS_PAGE : AUTH_PAGE;
+      return renderApp();
+    }
+
+    if (newPage === POSTS_PAGE) {
+      page = LOADING_PAGE;
+      renderApp();
+
+      return getPosts({ token: getToken() })
+        .then((newPosts) => {
+          page = POSTS_PAGE;
+          posts = newPosts;
+          renderApp();
+        })
+        .catch((error) => {
+          console.error(error);
+          goToPage(POSTS_PAGE);
+        });
+    }
+
+    if (newPage === USER_POSTS_PAGE) {
+      console.log(data.userId);
+
+      return getUserPosts({
+        token: getToken(),
+        id: data.userId
+      })
+        .then((userPosts) => {
+          page = USER_POSTS_PAGE;
+          posts = userPosts;
+          renderApp();
+        });
+    }
+
+    page = newPage;
+    renderApp();
+
+    return;
+  }
+
+  throw new Error("страницы не существует");
+};
+
+const renderApp = () => {
+  const appEl = document.getElementById("app");
+  if (page === LOADING_PAGE) {
+    return renderLoadingPageComponent({
+      appEl,
+      user,
+      goToPage,
+    });
+  }
+
+  if (page === AUTH_PAGE) {
+    return renderAuthPageComponent({
+      appEl,
+      setUser: (newUser) => {
+        user = newUser;
+        saveUserToLocalStorage(user);
+        goToPage(POSTS_PAGE);
+      },
+      user,
+      goToPage,
+    });
+  }
+
+  if (page === ADD_POSTS_PAGE) {
+    return renderAddPostPageComponent({
+      appEl,
+      onAddPostClick({ description, imageUrl }) {
+        addPost({
+          description: description,
+          imageUrl
+        })
+          .then(() => {
+            goToPage(POSTS_PAGE);
+          })
+      },
+    });
+  }
+
+  if (page === POSTS_PAGE || page === USER_POSTS_PAGE) {
+
+    return renderPostsPageComponent({
+      appEl,
+
+      onLikePostClick({ id, action, userId }) {
+        action({
+          token: getToken(),
+          id
+        })
+          .then(() => {
+            if (page === POSTS_PAGE) {
+              return getPosts({ token: getToken() });
+            } else {
+              return getUserPosts({ token: getToken(), id: userId });
+            }
+          })
+          .then((newPosts) => {
+            posts = newPosts;
+            renderApp();
+          })
+          .catch((error) => {
+            if (error.message === "Unauthorized") {
+              alert("Вы не авторизованы");
+            }
+          })
+      },
+
+      onDeleteButtonClick({ id, userId }) {
+        deletePost({
+          token: getToken(),
+          id
+        })
+          .then(() => {
+            if (page === POSTS_PAGE) {
+              return getPosts({ token: getToken() });
+            } else {
+              return getUserPosts({ token: getToken(), id: userId });
+            }
+          })
+          .then((newPosts) => {
+            posts = newPosts;
+            renderApp();
+          })
+          .catch((error) => {
+            if (error.message === "Unauthorized") {
+              alert("Вы не авторизованы");
+            }
+          })
+      }
+    })
+  }
+};
